Making search for a file on the manifest generic.
