+++ cibyl/sources/zuul/utils/artifacts/manifest.py
@@ -14,13 +14,14 @@
 #    under the License.
 """
 from dataclasses import dataclass, field
-from typing import List, NamedTuple, Optional, Sequence
+from typing import Iterable, List, NamedTuple, Optional, Sequence, Tuple
 
 from xsdata.formats.dataclass.parsers import JsonParser
 
 from cibyl.sources.zuul.apis import Artifact, ArtifactKind
 from cibyl.sources.zuul.apis.rest import ZuulBuildRESTClient as Build
 from cibyl.sources.zuul.utils.artifacts import ArtifactError
+from cibyl.utils.filtering import matches_regex
 from cibyl.utils.net import download_into_memory
 
 
@@ -179,3 +180,181 @@ class ManifestDownloader:
         downloader = ManifestDownloader.DownloadFromBuild(build, self.tools)
 
         return downloader.get_manifest()
+
+
+ManifestLevel = Iterable[ManifestItem]
+"""Represents a directory on the manifest."""
+
+
+class ManifestDigger:
+    """Tool used to transverse the directory tree of a manifest file. The
+    purpose of this tool is to be able to safely go down a branch on the
+    manifest's directory tree to its end. It allows to either observe the
+    contents of a level or to go deeper down the line. Once the end is
+    reached, this provides mechanism to be aware of so.
+    """
+
+    def __init__(self, manifest: Manifest):
+        """Constructor.
+
+        :param manifest: The manifest to dig through.
+        """
+        self._manifest = manifest
+        self._current_level = manifest.tree
+
+    @property
+    def manifest(self) -> Manifest:
+        """
+        :return: The manifest this is working with.
+        """
+        return self._manifest
+
+    @property
+    def current_level(self) -> ManifestLevel:
+        """
+        :return: Contents of the level the digger is currently at.
+        """
+        return self._current_level
+
+    def has_sublevel(self, name: str) -> bool:
+        """Checks whether there is a directory below this level with the
+        suggested name.
+
+        :param name: The name to check.
+        :return: True if there is, False if not.
+        """
+        return bool(self._get_sublevel(name))
+
+    def delve_into(self, level: str) -> None:
+        """Moves the digger down another level.
+
+        .. seealso::
+            :py:meth:`has_sublevel`
+                -> For checking if the level exists beforehand.
+
+        :param level: Name of the directory to go down to.
+        :raises ArtifactError: If the level does not exist.
+        """
+        if not self.has_sublevel(level):
+            msg = f"Manifest has no level: '{level}'."
+            raise ArtifactError(msg)
+
+        self._current_level = self._get_sublevel(level)
+
+    def _get_sublevel(self, level: str) -> Optional[ManifestLevel]:
+        """Gets the contents of a directory pending from the current level.
+
+        :param level: Name of the directory to search for.
+        :return: Contents of that directory. 'None' if it does not exist or
+            the name points to a file instead.
+        """
+        # Search for the sublevel
+        for item in self.current_level:
+            # Does the sublevel exist?
+            if item.name != level:
+                continue
+
+            # Check if at the end of the branch
+            if not item.children:
+                return None
+
+            # Not yet at the end -> Return the new level then
+            return item.children
+
+        return None
+
+
+class ManifestDiggerFactory:
+    """Factory for `ManifestDigger`.
+    """
+
+    def from_manifest(self, manifest: Manifest) -> ManifestDigger:
+        """Builds a new instance from a manifest file.
+
+        :param manifest: The manifest that the digger will work with.
+        :return: The digger's instance.
+        """
+        return ManifestDigger(manifest)
+
+
+class ManifestDir(str):
+    def __new__(cls, value: str) -> 'ManifestDir':
+        if not matches_regex(r'^(\/|(\/\w+)+)$', value):
+            raise ValueError(f"Invalid format for manifest path: '{value}'.")
+
+        return super().__new__(cls, value)
+
+
+class ManifestFile(str):
+    def __new__(cls, value: str) -> 'ManifestFile':
+        if not matches_regex(r'^(\/\w+)*(\/\w*\.\w*)$', value):
+            raise ValueError(f"Invalid format for manifest path: '{value}'.")
+
+        return super().__new__(cls, value)
+
+
+class ManifestFileSearch:
+    class SearchTerms(NamedTuple):
+        """Defines what the class has to search for.
+        """
+        paths: Iterable[str]
+        """Collection of paths where the desired file may be."""
+        files: Iterable[str]
+        """Collection of names the desired file may take."""
+
+    class Tools(NamedTuple):
+        diggers: ManifestDiggerFactory = ManifestDiggerFactory()
+
+    def __init__(self, tools: Tools = Tools()):
+        self._tools = tools
+
+    @property
+    def tools(self):
+        return self._tools
+
+    def find_in(
+        self,
+        manifest: Manifest,
+        search_terms: SearchTerms
+    ) -> Optional[Tuple[ManifestFile, ManifestItem]]:
+        for path in search_terms.paths:
+            level = self._get_manifest_level(manifest, ManifestDir(path))
+
+            if not level:
+                # The file's directory does not exist
+                continue
+
+            for file in search_terms.files:
+                # See if the file is at this level
+                results = next(
+                    (item for item in level if item.name == file),
+                    None
+                )
+
+                if not results:
+                    # The file is not here
+                    continue
+
+                # Generate path to the file
+                return ManifestFile(f'{path}{file}'), results
+
+        return None
+
+    def _get_manifest_level(
+        self,
+        manifest: Manifest,
+        path: ManifestDir
+    ) -> Optional[ManifestLevel]:
+        digger = self.tools.diggers.from_manifest(manifest)
+
+        for step in self._split_path(path):
+            if not digger.has_sublevel(step):
+                return None
+
+            digger.delve_into(step)
+
+        return digger.current_level
+
+    def _split_path(self, path: ManifestDir) -> Iterable[str]:
+        # Remove initial '/' to avoid empty splits
+        return path[1:].split('/')
